+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/buttons/ReplyForce.kt
@@ -15,8 +15,19 @@ data class ReplyForce(
     val forceReply: Boolean = true
 
     companion object {
-        val ReplyForceSelective = ReplyForce(true)
-        val ReplyForceNonSelective = ReplyForce(false)
+        fun Selective(inputFieldPlaceholder: String? = null) = ReplyForce(true, inputFieldPlaceholder)
+        fun NonSelective(inputFieldPlaceholder: String? = null) = ReplyForce(false, inputFieldPlaceholder)
+        val Selective = Selective()
+        val NonSelective = NonSelective()
+        val Default = ReplyForce()
+
+        @Deprecated("Renamed", ReplaceWith("ReplyForce.Selective"))
+        inline val ReplyForceSelective
+            get() = Selective
+        @Deprecated("Renamed", ReplaceWith("ReplyForce.NonSelective"))
+        inline val ReplyForceNonSelective
+            get() = NonSelective
+        @Deprecated("Renamed", ReplaceWith("ReplyForce.Default"))
         val ReplyForceDefault = ReplyForce()
     }